requirement.txt
Django==2.2.3
django-filter==2.2.0
djangorestframework==3.10.1
numpy==1.17.3
pandas==0.25.3
psycopg2==2.7.5
python-dateutil==2.8.1
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
